requirements.txt
progress
numpy
Pillow
tkinter
